# src/main/resources/swagger/api.yaml
openapi: 3.0.3
info:
    title: Scheduler API
    description: >-
        API REST para o MVP do SaaS de agendamento de serviços (MVP,single-tenant).
    version: 1.0.0
servers:
    - url: 'http://localhost:8080'
      description: Local
    - url: 'https://api.seusistema.com'
      description: Produção
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    responses:
        BadRequest:
            description: Requisição inválida
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        NotFound:
            description: Recurso não encontrado
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
...
        NoContent:
            description: Sem conteúdo
    schemas:
        ErrorResponse:
            type: object
            properties:
                message:
                    type: string
                status:
                    type: integer
                error:
                    type: string
                timestamp:
                    type: string
                    format: date-time
                    example: 09-12-2025
...
        AddressRequest:
            type: object
            required:
                - street
                - neighborhood
                - city
                - state
                - zipCode
            properties:
                street:
                    type: string
                    minLength: 1
                    x-field-extra-annotation: >
                        @jakarta.validation.constraints.NotBlank(message = "Rua é
                        obrigatória")
                number:
                    type: string
                complement:
                    type: string
                neighborhood:
                    type: string
                    minLength: 1
                    x-field-extra-annotation: >
                        @jakarta.validation.constraints.NotBlank(message = "Bairro é
                        obrigatório")
                city:
                    type: string
                    minLength: 1
                    x-field-extra-annotation: >
                        @jakarta.validation.constraints.NotBlank(message = "Cidade é
                        obrigatória")
                state:
                    type: string
                    pattern: '^[A-Z]{2}$'
                    x-field-extra-annotation: |
                        @jakarta.validation.constraints.Pattern(
                          regexp = "^[A-Z]{2}$",
                          message = "Estado deve conter 2 letras (ex: SP)"
                        )
                zipCode:
                    type: string
                    pattern: '\d{5}-\d{3}'
                    x-field-extra-annotation: |
                        @jakarta.validation.constraints.Pattern(
                          regexp = "\\d{5}-\\d{3}",
                          message = "CEP inválido (ex: 01310-100)"
                        )
        AddressResponse:
            type: object
            properties:
                street:
                    type: string
                number:
                    type: string
                complement:
                    type: string
                neighborhood:
                    type: string
                city:
                    type: string
                state:
                    type: string
                zipCode:
                    type: string
        ClientResponse:
            type: object
            required:
                - id
                - name
                - email
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
                email:
                    type: string
                    format: email
                phone:
                    type: string
                address:
                    $ref: '#/components/schemas/AddressResponse'
                createdAt:
                    type: string
                    format: date-time
                    example: 09-12-2025
                updatedAt:
                    type: string
                    format: date-time
                    example: 09-12-2025
        ClientRequest:
            type: object
            required:
                - name
                - email
            properties:
                name:
                    type: string
                    minLength: 1
                email:
                    type: string
                    format: email
                phone:
                    type: string
                    pattern: '\d{10,11}'
                address:
                    $ref: '#/components/schemas/AddressRequest'
        ClientUpdate:
            type: object
            properties:
                name:
                    type: string
                    minLength: 1
                email:
                    type: string
                    format: email
                phone:
                    type: string
                    pattern: '\d{10,11}'
                address:
                    $ref: '#/components/schemas/AddressRequest'
        ProfessionalResponse:
            type: object
            required:
                - id
                - name
            properties:
                id:
                    type: integer
                    format: int64
                    example: 1
                name:
                    type: string
                    example: Joana Pereira
                bio:
                    type: string
                    example: Especialista em alongamento e nail art
                phone:
                    type: string
                    example: +55 51 98888-1111
                email:
                    type: string
                    format: email
                    example: joana.pereira@email.com
                address:
                    $ref: '#/components/schemas/AddressResponse'
                specialtyIds:
                    type: array
                    items:
                        type: integer
                        format: int64
                    example:
                        - 1
                        - 3
                        - 5
                rating:
                    type: number
                    format: decimal
                    example: 4.8
                totalReviews:
                    type: integer
                    example: 152
                specialization:
                    type: string
                    nullable: true
                    example: Especialista em Manicure Basica
                    description: Especializacao principal do profissional (derivada do primeiro specialtyId)
                workStartTime:
                    type: string
                    format: time
                    example: '09:00'
                workEndTime:
                    type: string
                    format: time
                    example: '19:00'
                workingDays:
                    type: array
                    items:
                        type: string
                    example:
                        - MONDAY
                        - TUESDAY
                        - WEDNESDAY
                        - THURSDAY
                        - FRIDAY
                active:
                    type: boolean
                    example: true
                createdAt:
                    type: string
                    format: date-time
                    example: '2025-12-15T15:55:16'
                updatedAt:
                    type: string
                    format: date-time
                    example: '2025-12-15T15:55:16'
        ProfessionalRequest:
            type: object
            required:
                - name
                - email
                - phone
                - workStartTime
                - workEndTime
                - workingDays
                - active
            properties:
                name:
                    type: string
                email:
                    type: string
                    format: email
                phone:
                    type: string
                address:
                    $ref: '#/components/schemas/AddressRequest'
                bio:
                    type: string
                specialtyIds:
                    type: array
                    items:
                        type: integer
                        format: int64
                workStartTime:
                    type: string
                    format: time
                    example: '09:00'
                workEndTime:
                    type: string
                    format: time
                    example: '19:00'
                workingDays:
                    type: array
                    items:
                        type: string
                    example:
                        - MONDAY
                        - TUESDAY
                        - WEDNESDAY
                        - THURSDAY
                        - FRIDAY
                active:
                    type: boolean
        ReviewRequest:
            type: object
            required:
                - rating
            properties:
                rating:
                    type: integer
                    minimum: 1
                    maximum: 5
                comment:
                    type: string
                    maxLength: 500
        ReviewResponse:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                professionalId:
                    type: integer
                    format: int64
                rating:
                    type: integer
                comment:
                    type: string
                createdAt:
                    type: string
                    format: date-time
        AvailabilityResponse:
            type: object
            required:
                - professionalId
                - date
                - durationMinutes
                - availableSlots
            properties:
                professionalId:
                    type: integer
                    format: int64
                    example: 1
                    description: ID do profissional
                date:
                    type: string
                    format: date
                    example: '2024-04-25'
                    description: Data consultada
                durationMinutes:
                    type: integer
                    example: 30
                    description: Duração em minutos considerada para os slots
                availableSlots:
                    type: array
                    items:
                        type: string
                        format: time
                        pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
                    example:
                        - '09:00'
                        - '09:30'
                        - '10:00'
                        - '10:30'
                        - '14:00'
                    description: Lista de horarios disponiveis no formato HH:mm
...
security:
    - bearerAuth: []
paths:
    /clients/me:
        get:
            tags:
                - clients
            summary: Obter perfil do cliente logado
            operationId: getMyClientProfile
            responses:
                '200':
                    description: Perfil do cliente
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ClientResponse'
        put:
            tags:
                - clients
            summary: Atualizar perfil do cliente logado
            operationId: updateMyClientProfile
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ClientRequest'
            responses:
                '200':
                    description: Perfil atualizado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ClientResponse'
    /clients:
        get:
            tags:
                - clients
            summary: Listar clientes
            operationId: listClients
            responses:
                '200':
                    description: Lista de clientes
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ClientResponse'
        post:
            tags:
                - clients
            summary: Criar cliente
            operationId: createClient
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ClientRequest'
            responses:
                '201':
                    description: Cliente criado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ClientResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
    '/clients/{clientId}':
        get:
            tags:
                - clients
            summary: Buscar cliente por ID
            operationId: getClient
            parameters:
                - name: clientId
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: Cliente encontrado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ClientResponse'
                '404':
                    $ref: '#/components/responses/NotFound'
        patch:
            tags:
                - clients
            summary: Atualizar cliente
            operationId: updateClient
            parameters:
                - name: clientId
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ClientUpdate'
            responses:
                '200':
                    description: Cliente atualizado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ClientResponse'
        delete:
            tags:
                - clients
            summary: Remover cliente (soft delete recomendado)
            operationId: deleteClient
            parameters:
                - name: clientId
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                '204':
                    $ref: '#/components/responses/NoContent'
    /clients/search:
        get:
            tags:
                - clients
            summary: Buscar clientes por Nome (Partial Match)
            description: Pesquisa por nome parcial.
            operationId: getClientByName
            parameters:
                - name: name
                  in: query
                  required: true
                  schema:
                      type: string
                  description: Nome ou parte do nome para pesquisa
            responses:
                '200':
                    description: Lista de clientes encontrados
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ClientResponse'
    /professionals/me:
        get:
            tags:
                - professionals
            summary: Obter perfil do profissional logado
            operationId: getMyProfessionalProfile
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Perfil do profissional
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProfessionalResponse'
                '404':
                    $ref: '#/components/responses/NotFound'
    /professionals:
        get:
            tags:
                - professionals
            summary: Listar profissionais
            operationId: listProfessionals
            parameters:
                - name: departmentId
                  in: query
                  required: false
                  schema:
                      type: integer
                      format: int64
                  description: |
                    Filtrar por ID do departamento/serviço (departmentId).
                    Nota: O parâmetro é chamado 'departmentId' mas pode ser referenciado como 'serviceId' 
                    no frontend - ambos representam o mesmo conceito.
            responses:
                '200':
                    description: Lista de profissionais
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ProfessionalResponse'
        post:
            tags:
                - professionals
            summary: Criar profissional
            operationId: createProfessional
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ProfessionalRequest'
            responses:
                '201':
                    description: Profissional criado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProfessionalResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
    '/professionals/{professionalId}':
        get:
            tags:
                - professionals
            summary: Buscar profissional por ID
            operationId: getProfessional
            parameters:
                - name: professionalId
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: Profissional encontrado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProfessionalResponse'
                '404':
                    $ref: '#/components/responses/NotFound'
        put:
            tags:
                - professionals
            summary: Atualizar profissional
            operationId: updateProfessional
            parameters:
                - name: professionalId
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ProfessionalRequest'
            responses:
                '200':
                    description: Profissional atualizado
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProfessionalResponse'
        delete:
            tags:
                - professionals
            summary: Remover profissional
            operationId: deleteProfessional
            parameters:
                - name: professionalId
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                '204':
                    $ref: '#/components/responses/NoContent'
    '/professionals/{professionalId}/reviews':
        post:
            tags:
                - professionals
            summary: Avaliar profissional
            operationId: createReview
            parameters:
                - name: professionalId
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ReviewRequest'
            responses:
                '201':
                    description: Avaliação criada
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ReviewResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
    '/professionals/{professionalId}/availability':
        get:
            tags:
                - professionals
            summary: Consultar disponibilidade de horarios de um profissional
            description: |
                Retorna lista de slots de horarios disponiveis para agendamento em uma data especifica.
                Considera:
                - Horário de trabalho do profissional (workStartTime, workEndTime)
                - Dias de trabalho (workingDays)
                - Agendamentos já existentes (confirmados ou pendentes)
                - Duração do serviço (durationMinutes)
            operationId: getProfessionalAvailability
            parameters:
                - name: professionalId
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
                  description: ID do profissional
                - name: date
                  in: query
                  required: true
                  schema:
                      type: string
                      format: date
                      example: '2024-04-25'
                  description: Data para consulta (formato YYYY-MM-DD)
                - name: durationMinutes
                  in: query
                  required: false
                  schema:
                      type: integer
                      minimum: 15
                      maximum: 480
                      default: 30
                      example: 30
                  description: 'Duracao do servico em minutos (padrao: 30)'
            responses:
                '200':
                    description: Lista de horarios disponiveis
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AvailabilityResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '404':
                    $ref: '#/components/responses/NotFound'
...